fix bug

+++ src/main/java/transaction/PopularItemTransaction.java
@@ -1,5 +1,6 @@
 package transaction;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Date;
@@ -94,7 +95,7 @@ public class PopularItemTransaction extends AbstractTransaction {
 
             long maxQuantity = executeQuery(PreparedQueries.getMaxOLQuantity, orderId, districtId, warehouseId)
                     .get(0).getBigDecimal(0).longValue();
-            List<Row> getPopularItemIdsResult = executeQuery(PreparedQueries.getPopularItems, orderId, districtId, warehouseId, maxQuantity);
+            List<Row> getPopularItemIdsResult = executeQuery(PreparedQueries.getPopularItems, orderId, districtId, warehouseId, BigDecimal.valueOf(maxQuantity));
 
             // builder.append(String.format("max quantity: %d", maxQuantity));
